# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VELOXA API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * @nullable
 */
export type OwnerSocialLinks = { [key: string]: unknown } | null;

export interface Owner {
  id: number;
  username: string;
  name: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  socialLinks?: OwnerSocialLinks;
  createdAt?: string;
}

/**
 * @nullable
 */
export type OwnerWithCardSocialLinks = { [key: string]: unknown } | null;

export interface OwnerWithCard {
  id: number;
  username: string;
  name: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  socialLinks?: OwnerWithCardSocialLinks;
  nfcToken: string;
  profileUrl: string;
}

export interface OwnerInput {
  name: string;
  username: string;
  title?: string;
  company?: string;
  bio?: string;
  phone?: string;
  email?: string;
  website?: string;
  city?: string;
  industry?: string;
}

export type OwnerUpdateSocialLinks = { [key: string]: unknown };

export interface OwnerUpdate {
  name?: string;
  title?: string;
  company?: string;
  bio?: string;
  avatarUrl?: string;
  phone?: string;
  email?: string;
  website?: string;
  instagram?: string;
  linkedin?: string;
  city?: string;
  industry?: string;
  socialLinks?: OwnerUpdateSocialLinks;
}

export interface Sponsor {
  id: number;
  name: string;
  tagline: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bgImageUrl?: string | null;
  /** @nullable */
  teaserImageUrl?: string | null;
  ctaText: string;
  ctaUrl: string;
  /** @nullable */
  targetCategories?: string[] | null;
}

export interface SponsorInput {
  name: string;
  tagline: string;
  logoUrl?: string;
  bgImageUrl?: string;
  teaserImageUrl?: string;
  ctaText: string;
  ctaUrl: string;
  targetCategories?: string[];
}

export interface SponsorLead {
  id: number;
  sponsorId: number;
  ownerId: number;
  /** @nullable */
  nfcToken?: string | null;
  /** @nullable */
  sourceUsername?: string | null;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  status: string;
  createdAt: string;
}

export interface SponsorLeadInput {
  sponsorId: number;
  ownerId: number;
  nfcToken?: string;
  sourceUsername?: string;
  name: string;
  email: string;
  phone?: string;
}

export interface SponsorLeadDetail {
  id: number;
  sponsorId: number;
  ownerId: number;
  /** @nullable */
  nfcToken?: string | null;
  /** @nullable */
  sourceUsername?: string | null;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  status: string;
  createdAt: string;
  /** @nullable */
  sponsorName?: string | null;
  /** @nullable */
  ownerName?: string | null;
}

export interface LeadStatusUpdate {
  status: string;
}

export interface NfcCardDetail {
  id: number;
  /** @nullable */
  ownerId?: number | null;
  token: string;
  status: string;
  /** @nullable */
  activatedAt?: string | null;
  createdAt: string;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  ownerUsername?: string | null;
  /** @nullable */
  profileUrl?: string | null;
}

export interface CardStatusUpdate {
  status: string;
}

export interface CardStatus {
  token: string;
  status: string;
  /** @nullable */
  ownerId?: number | null;
}

export interface ActivationInput {
  name: string;
  company?: string;
  title?: string;
  phone?: string;
  email: string;
  website?: string;
  instagram?: string;
  linkedin?: string;
  city?: string;
  industry?: string;
  bio?: string;
  avatarUrl?: string;
  password: string;
}

export interface ActivationResult {
  ownerId: number;
  token: string;
  profileUrl: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResult {
  ownerId: number;
  name: string;
  email: string;
}

export interface CrmContact {
  id: number;
  /** @nullable */
  ownerId?: number | null;
  name: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  industry?: string | null;
  source: string;
  /** @nullable */
  nfcToken?: string | null;
  createdAt: string;
}

export type AdminStatsLeadsByStatusItem = {
  status: string;
  count: number;
};

export interface AdminStats {
  totalLeads: number;
  totalOwners: number;
  totalSponsors: number;
  newLeadsToday: number;
  leadsByStatus?: AdminStatsLeadsByStatusItem[];
}

export type MatchSponsorParams = {
ownerId: number;
};

export type AdminGetLeadsParams = {
status?: string;
sponsorId?: number;
ownerId?: number;
};