# lib/api-zod/src/generated/types/activationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VELOXA API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivationInput {
  name: string;
  company?: string;
  title?: string;
  phone?: string;
  email: string;
  website?: string;
  instagram?: string;
  linkedin?: string;
  city?: string;
  industry?: string;
  bio?: string;
  avatarUrl?: string;
  password: string;
}