# lib/api-zod/src/generated/types/activationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VELOXA API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivationResult {
  ownerId: number;
  token: string;
  profileUrl: string;
}