# lib/api-zod/src/generated/types/cardStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VELOXA API
 * OpenAPI spec version: 0.1.0
 */

export interface CardStatus {
  token: string;
  status: string;
  /** @nullable */
  ownerId?: number | null;
}