# lib/api-zod/src/generated/types/nfcCardDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VELOXA API
 * OpenAPI spec version: 0.1.0
 */

export interface NfcCardDetail {
  id: number;
  /** @nullable */
  ownerId?: number | null;
  token: string;
  status: string;
  /** @nullable */
  activatedAt?: string | null;
  createdAt: string;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  ownerUsername?: string | null;
  /** @nullable */
  profileUrl?: string | null;
}